refactor(core): one spelling of "truthy", and a guard that keeps it that way
debug_switches::is_truthy documents itself as "the spelling every switch in
the tree already accepts, gathered here so a ninth one cannot quietly accept a
different set". That was false. Five other sites spelled out the same
matches! — two of them in this very crate — and is_truthy itself had no caller
outside its own module. Nothing failed while they agreed, because each copy
was locally correct; the cost only arrives when one of them changes.

The four `GGLIB_DISABLE_*` readers now call a new debug_switches::enabled,
which reads the live environment on every call. Caching would ignore a .env
loaded after startup and would leak between tests that set one switch per
case. gglib-runtime's is_truthy_flag keeps its name — it is that crate's
vocabulary — and delegates.

A unit test cannot see a copy nobody has written yet, so the guard is a grep:
one test walks the crate tree for the literal set and fails on any occurrence
outside this module. Verified by adding one back, which failed it.

The tests move to debug_switches_tests.rs, per the models_tests.rs convention
— the module is a list of constants and three small functions, and the tests
are now longer than all of it.
